+++ saiph/reduction/famd_test.py
@@ -29,16 +29,16 @@ def test_fit_mix() -> None:
             "Dim. 2": [0.0, 0.0],
         }
     )
-    expected_v: NDArray[float] = np.array(
+    expected_v: NDArray[np.float_] = np.array(
         [
             [0.57735, 0.408248, -0.408248, -0.408248, 0.408248],
             [0.816497, -0.288675, 0.288675, 0.288675, -0.288675],
         ]
     )
-    expected_s: NDArray[float] = np.array([1.224745e00, 0.0])
-    expected_u: NDArray[float] = np.array([[-1.0, 1.0], [1.0, 1.0]])
-    expected_explained_var: NDArray[float] = np.array([1.5, 0.0])
-    expected_explained_var_ratio: NDArray[float] = np.array([1.0, 0.0])
+    expected_s: NDArray[np.float_] = np.array([1.224745e00, 0.0])
+    expected_u: NDArray[np.float_] = np.array([[-1.0, 1.0], [1.0, 1.0]])
+    expected_explained_var: NDArray[np.float_] = np.array([1.5, 0.0])
+    expected_explained_var_ratio: NDArray[np.float_] = np.array([1.0, 0.0])
 
     assert_frame_equal(result, expected_result, check_exact=False, atol=0.01)
     assert_allclose(model.V, expected_v, atol=0.01)
@@ -48,7 +48,7 @@ def test_fit_mix() -> None:
     assert_allclose(model.explained_var_ratio, expected_explained_var_ratio, atol=0.01),
     assert_allclose(model.variable_coord, model.V.T)
     assert np.array_equal(
-        model._modalities,
+        model._modalities,  # type: ignore
         [
             f"tool{DUMMIES_PREFIX_SEP}hammer",
             f"tool{DUMMIES_PREFIX_SEP}toaster",
@@ -70,7 +70,7 @@ def test_fit_mix() -> None:
         atol=0.01,
     )
     assert np.array_equal(
-        model._modalities,
+        model._modalities,  # type: ignore
         [
             f"tool{DUMMIES_PREFIX_SEP}hammer",
             f"tool{DUMMIES_PREFIX_SEP}toaster",
